resolve conflict and apply changes for new usart
