+++ boards/nucleo_f429zi/Makefile
@@ -21,9 +21,11 @@ flash-debug: target/$(TARGET)/debug/$(PLATFORM).elf
 flash: target/$(TARGET)/release/$(PLATFORM).elf
 	sudo $(OPENOCD) $(OPENOCD_OPTIONS) -c "init; reset halt; flash write_image erase $(CWD)/$<; verify_image $(CWD)/$<; reset; shutdown"
 
-APP=../../../libtock-c/examples/tests/console_timeout/build/cortex-m4/cortex-m4.tbf
+APP=../../../libtock-c/examples/tests/hello_loop/build/cortex-m4/cortex-m4.tbf
+# APP=../../../libtock-c/examples/tests/console/build/cortex-m4/cortex-m4.tbf
+# APP=../../../libtock-c/examples/tests/console_timeout/build/cortex-m4/cortex-m4.tbf
 # APP=../../../libtock-c/examples/buttons/build/cortex-m4/cortex-m4.tbf
-# APP=../../../libtock-c/examples/buttons/build/cortex-m4/cortex-m4.tbf
+# APP=../../../libtock-c/examples/blink/build/cortex-m4/cortex-m4.tbf
 KERNEL=$(CWD)/target/$(TARGET)/debug/$(PLATFORM).elf
 KERNEL_WITH_APP=$(CWD)/target/$(TARGET)/debug/$(PLATFORM)-app.elf
